ats-integration-microservice/pdf-analysis-project/requirements.txt
click==8.1.8
joblib==1.5.2
nltk==3.9.1
psycopg2-binary==2.9.10
PyMuPDF==1.26.4
python-dotenv==1.1.1
regex==2025.9.18
tqdm==4.67.1
psutil==5.9.8
pytest==7.4.4
flask==3.0.0
requests==2.31.0
